refactoring

+++ instrument_constants.py
@@ -7,7 +7,7 @@ class Constants:
     """All of the constants that are needed."""
 
     iroa_d: float = -46.25
-    offx: float = 4240
-    offy: float = 4240
-    nx_num: float = 8800
-    ny_num: float = 8800
+    offx: int = 4240
+    offy: int = 4240
+    nx_num: int = 8800
+    ny_num: int = 8800
